replacements: adjust for com.endlessm.Installer change
I've modified eos-installer to install its desktop file to the usual
location, `/usr/share/applications`, but to set the `NoDisplay=true`
key. On a live boot, this file is patched to remove that key.

This means that eos-gates cannot use the mere presence of the
com.endlessm.Installer.desktop file as an indication that we're on a
live system. Instead, check that it exists *and* that it does not have
`NoDisplay=true` set.

https://phabricator.endlessm.com/T17082