Stop using endlessm-app:// scheme to launch Software
In previous versions of Endless OS, our gnome-shell was patched to
implement an endlessm-app:// URI scheme that would launch the given
application with startup notification. I believe this was important in
the past because of the eos-speedwagon splash screen that used to be
shown during the startup-notification phase, but we removed that long
ago.

This is the only remaining user of that URI scheme, and it's only used
to launch GNOME Software when no match is found. Other apps are
launched via Gio.DesktopAppInfo.launch().

Instead, activate GNOME Software over D-Bus, instructing it to show
the overview.

https://phabricator.endlessm.com/T31635